cmake: adapt for VS
Add header files to sources to make them part of the project when
opening in Visual Studio.

Correct MSVC compiler flags in order to compile with coroutines.

+++ lib/core/src/sign/commands.cpp
@@ -18,7 +18,7 @@ void WriteToBuffer(T i, char * from, char * to) noexcept
 
 void WriteToBuffer(std::string_view sv, char * from, char * to) noexcept
 {
-   assert(sv.size() < to - from);
+   assert(static_cast<int>(sv.size()) < to - from);
    *from++ = static_cast<unsigned char>(sv.size());
    std::memcpy(from, sv.data(), sv.size());
    (void)to;
